% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcpmMinusloglik.R
\name{lcpmMinusloglik}
\alias{lcpmMinusloglik}
\title{The Negative Log-Likelihood Function for a Log Cumulative Probability Model}
\usage{
lcpmMinusloglik(betapar, Xa1, XaJ, Xaj1, Xaj2)
}
\arguments{
\item{betapar}{a vector of values.}

\item{Xa1}{matrix of covariates for all subjects with the lowest ordinal outcome value 1.}

\item{XaJ}{matrix of covariates for all subjects with the largest ordinal outcome value J.}

\item{Xaj1}{matrix of covariates for all subjects with the ordinal outcomes with value 1 < j < J.}

\item{Xaj2}{matrix of covariates for all subjects with the ordinal outcome with value 1 < j < J but lagged by 1.}
}
\value{
value of the negative log-likelihood evaluated at betapar
}
\description{
\code{lcpmMinusloglik} provides the negative of the log-likelihood function for a Generalized Linear Model with a log link and ordinal outcomes to be minimized in functions \code{\link{lcpm}} and \code{\link{ppm}}.
}
